Start runtime fix
